requirements.txt
aiohttp==3.7.4.post0
alembic==1.7.4
amqp==5.0.6
async-timeout==3.0.1
asyncpg==0.24.0
asyncpgsa==0.27.1
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.2
chardet==4.0.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
greenlet==1.1.2
idna==3.3
kombu==5.1.0
Mako==1.1.5
MarkupSafe==2.0.1
multidict==5.2.0
prompt-toolkit==3.0.21
psycopg2-binary==2.9.1
python-dateutil==2.8.2
python-environ==0.4.54
pytz==2021.3
six==1.16.0
SQLAlchemy==1.3.24
typing-extensions==3.10.0.2
vine==5.0.0
wcwidth==0.2.5
yarl==1.7.0
